+++ src/commands/deploy.ts
@@ -1,9 +1,9 @@
-import { Config, Project } from './common/Config';
-import { ScreenPrinter } from './console/ScreenPrinter';
-import { awaitPipelineCompletion, getPipeline, IPipeline } from './common/pipelines';
-import { findJob, playJob, StatusCode } from './common/api';
-import { IJob } from './common/iJob';
-import { Yargs } from './common/Yargs';
+import { Config, Project } from '../common/Config';
+import { ScreenPrinter } from '../console/ScreenPrinter';
+import { awaitPipelineCompletion, getPipeline, IPipeline } from '../common/pipelines';
+import { findJob, playJob, StatusCode } from '../common/api';
+import { IJob } from '../common/iJob';
+import { Yargs } from '../common/Yargs';
 
 export function runDeploy(args) {
   const yargs = new Yargs(args);
@@ -13,11 +13,11 @@ export function runDeploy(args) {
   const promises = config.projects.map(async function(project) {
     screenPrinter.addProject(project);
     screenPrinter.print();
-    const pipeline = await getPipeline(project, config, yargs.ref, screenPrinter) as IPipeline;
-    if (!pipeline.id) return;
-    const job = (await getJob(project, pipeline, config, screenPrinter)) as IJob;
+    const pipeline = await getPipeline(project, config, yargs.ref, screenPrinter);
+    if (pipeline.status !== StatusCode.Success || !pipeline.data.id) return;
+    const job = (await getJob(project, pipeline.data, config, screenPrinter)) as IJob;
     if (!job.id) return;
-    const trigJob = await triggerJob(project, job, config, screenPrinter) as IJob;
+    const trigJob = (await triggerJob(project, job, config, screenPrinter)) as IJob;
     if (!trigJob.id) return;
     return awaitComplete(project, config, yargs, screenPrinter);
   });
@@ -60,9 +60,14 @@ export async function getJob(
   );
 }
 
-export async function triggerJob(project: Project, job: IJob, config: Config, screenPrinter: ScreenPrinter) {
+export async function triggerJob(
+  project: Project,
+  job: IJob,
+  config: Config,
+  screenPrinter: ScreenPrinter,
+) {
   return playJob(config.uri, project.id, job.id).then(
-    (data) => {
+    data => {
       screenPrinter.setProjectSuccess(project, 'IJob Started');
       return data;
     },