+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Change log format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## unreleased
 ### Changed
 - number of checked pipelines when searching for given ref
+- pipelines now return actual status
 ### Fixed
 - trigger job api
 - yargs help print now detects tty width